modified the auth controller file structure
