+++ server/controllers/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    AuthController: require('./auth')
+}
